// src/types/interfaces.d.ts
export type SessionTypesAsStrings = 'softphone' | 'screenShare' | 'screenRecording' | 'liveScreenMonitoring' | 'collaborateVideo' | 'unknown';
export declare enum SessionTypes {
    softphone = "softphone",
    collaborateVideo = "collaborateVideo",
    acdScreenShare = "screenShare",
    screenRecording = "screenRecording",
    liveScreenMonitoring = "liveScreenMonitoring",
    unknown = "unknown"
}
export interface ISessionInfo extends IPendingSession {
}
export interface IPendingSession {
    sessionId: string;
    id: string;
    autoAnswer: boolean;
    toJid: string;
    fromJid: string;
    conversationId: string;
    originalRoomJid?: string;
    sdpOverXmpp?: boolean;
    privAnswerMode?: 'Auto';
    fromUserId?: string;
    roomJid?: string;
    accepted?: boolean;
    meetingId?: string;
    sessionType: SessionTypes | SessionTypesAsStrings;
}